niffler_tests/requirements.txt
python-dotenv~=1.0.1
pytest~=8.2.2
requests==2.32.4
Faker~=37.4.0
selenium~=4.33.0
webdriver-manager~=4.0.2